Email/Linking/Account
When the user clicks on buy/sell now they will be either given a link or
be asked to enter an email.

When an email is entered, it will be sent to the poster. This is mostly
implemented in the two partial views for buy/sell. All of the jQuery is
within a section in the Index view for the market controller.

I also linked the BUY and SELL buttons on the home page.